rail-recommender/data-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
psycopg2-binary==2.9.9
python-dotenv==0.19.0